pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spotify-downloader"
version = "1.0.1"
description = "Download Spotify playlists, albums, and tracks in FLAC/MP3 format from free sources"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Mokshit Bindal", email = "your-email@example.com"}
]
keywords = ["spotify", "downloader", "music", "flac", "mp3", "audio"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Sound/Audio",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "spotipy>=2.24.0",
    "yt-dlp>=2024.10.7",
    "mutagen>=1.47.0",
    "pydub>=0.25.1",
    "click>=8.1.7",
    "rich>=13.7.0",
    "requests>=2.31.0",
    "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/MokshitBindal/Spotify_Downloader"
Documentation = "https://github.com/MokshitBindal/Spotify_Downloader#readme"
Repository = "https://github.com/MokshitBindal/Spotify_Downloader"
"Bug Tracker" = "https://github.com/MokshitBindal/Spotify_Downloader/issues"

[project.scripts]
spotify-downloader = "main:main"
spotdl = "main:main"

[tool.setuptools]
py-modules = ["main"]
packages = ["src"]
include-package-data = true
